refactor: Create new file StatusSettingsHelpers.ts
This will enable me to move a lot of the status settings
code out of SettingsTab.ts - to enable the code to be testable,
and eliminate a lot of the complexity and deep nesting in SettingsTab.ts.

Longer term, it should allow me to allow users to edit some
aspects of the pre-built statuses, like DONE and IN_PROGRESS.
